Correct type - recast
git-svn-id: file:///home/jj/hercules.svn/trunk@6747 956126f8-22a0-4046-8f4a-272fa8102e63

+++ archlvl.c
@@ -278,7 +278,7 @@ void force_facbit(int bitno, int enable, BYTE mode)
 {
 int fbyte, fbit;
 char buf[10];
-#define i2a(_int) ( (snprintf(buf,sizeof(buf),"%d", _int) <= sizeof(buf)) ? buf : "?" )
+#define i2a(_int) ( (snprintf(buf,sizeof(buf),"%d", _int) <= (int)sizeof(buf)) ? buf : "?" )
 
     fbyte = bitno / 8;
     fbit = 0x80 >> (bitno % 8);
